change vmreqmanager to module

+++ src/Virtualization/VirtualMachines/VMRequestManager/VMRequestManager.h
@@ -23,7 +23,7 @@ protected:
 
     /** Bool that control if the main queue does not receive messages */
         bool schedulerQueueBlocked;
-        int dockermem;
+        double  dockermem;
         string id;
         vector <RunningContainer*> rContainer;
     /** If schedulerQueueBlocked = false, the requests are allocated in this queue */
